+++ CHANGES.md
@@ -13,11 +13,13 @@ Release Notes.
 * Support [go-redis](https://github.com/redis/go-redis) v9 redis client framework.
 * Support collecting [Native HTTP](https://pkg.go.dev/net/http) URI parameter on server side.
 * Support [Mongo](https://github.com/mongodb/mongo-go-driver) database client framework.
+* Support [Native SQL](https://pkg.go.dev/net/http) database client framework with [MySQL Driver](github.com/go-sql-driver/mysql).
 
 #### Documentation
 
 #### Bug Fixes
 * Fix throw panic when log the tracing context before agent core initialized.
+* Fix plugin version matcher `tryToFindThePluginVersion` to support capital letters in module paths and versions.
 
 #### Issues and PR
 - All issues are [here](https://github.com/apache/skywalking/milestone/180?closed=1)